pyproject.toml
[project]
name = "WrittenBookEditor"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "XiYang6666", email = "1782356858@qq.com" }]
dependencies = [
    "PySide6>=6.7.3",
    "pillow>=10.4.0",
    "numpy>=2.1.1",
    "nbtlib>=2.0.4",
    "chardet>=5.2.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = { text = "MIT" }

[tool.pdm.scripts]
generate_ui = { shell = """
pyside6-uic src/writtenbookeditor/gui/mainwindow.ui -o src/writtenbookeditor/gui/ui_mainwindow.py &&\
pyside6-uic src/writtenbookeditor/gui/jump_form.ui -o src/writtenbookeditor/gui/ui_jump_form.py &&\
pyside6-uic src/writtenbookeditor/gui/encoding_form.ui -o src/writtenbookeditor/gui/ui_encoding_form.py\
""" }
dev = "python -m src.writtenbookeditor"
build = "pyinstaller ./main.spec -y "

[tool.pdm.dev-dependencies]
dev = ["pytest>=8.3.3"]
[tool.pdm]
distribution = false

[tool.black]
line-length = 200

# [tool.pyright]
# exclude = ["**/ui_*.py"]
